Refactored FhirPathFunctions to its own class. Added better kdocs and renamed some variables.
